+++ examples/nuxt-app/test/features/media/media.ts
@@ -18,7 +18,11 @@ Then('the media page should have the timestamp of {string}', (date: string) => {
 Then(
   'the media page should include a video embed with source {string}',
   (url: string) => {
-    cy.get('.tide-media__embedded-video').should('have.attr', 'src', url)
+    cy.get('.tide-media__embedded-video .rpl-media-embed__video').should(
+      'have.attr',
+      'src',
+      url
+    )
   }
 )